+++ src/DataProvider.php
@@ -16,6 +16,7 @@ final class DataProvider {
   use \Nette\SmartObject;
 
   public const ANNOTATION_NAME = "data";
+  public const PROVIDER_ANNOTATION_NAME = "dataProvider";
 
   private Reader $annotationsReader;
 
@@ -23,10 +24,30 @@ final class DataProvider {
     $this->annotationsReader = $annotationsReader;
   }
 
-  public function getData(string $class, string $method): array {
+  /**
+   * @throws InvalidDataProviderException
+   */
+  public function getData(object $class, string $method): array {
     $reflection = new ReflectionMethod($class, $method);
     if($reflection->getNumberOfParameters() < 1) {
       return [];
+    }
+    $dataProvider = $this->annotationsReader->getAnnotation(static::PROVIDER_ANNOTATION_NAME, $class, $method);
+    if(is_string($dataProvider)) {
+      $className = $reflection->getDeclaringClass()->getName();
+      try {
+        $reflection = new ReflectionMethod($class, $dataProvider);
+        if(!$reflection->isPublic()) {
+          throw new InvalidDataProviderException("Method $className::$dataProvider is not public.");
+        }
+        $result = call_user_func([$class, $dataProvider]);
+        if(!is_array($result)) {
+          throw new InvalidDataProviderException("Method $className::$dataProvider has to return an array.");
+        }
+        return $result;
+      } catch(\ReflectionException $e) {
+        throw new InvalidDataProviderException("Method $className::$dataProvider does not exist.", 0, $e);
+      }
     }
     /** @var mixed $value */
     $value = $this->annotationsReader->getAnnotation(static::ANNOTATION_NAME, $class, $method);